pyproject.toml
[project]
name = "config-loader"
version = "0.1.0"
description = "A Python package for loading and validating configuration from YAML files."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pytest-cov>=6.2.1",
]
